backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Document Processing
pdfplumber==0.10.3
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1

# NLP & ML
scikit-learn==1.4.0
numpy==1.26.3
google-generativeai==0.3.2

# Email
resend==0.7.0

# HTTP Client
httpx==0.26.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
